+++ source/EasyDev.py
@@ -6,6 +6,7 @@ from com.sun.star.lang import XServiceInfo
 from easydev.debug import Debug
 from easydev.tools import Tools, Arrays
 from easydev.email import EMail
+from easydev.lochart import LOChart
 from easydev.localc import LOCalc
 from easydev.lowriter import LOWriter
 from easydev.dialog import LODialog
@@ -14,7 +15,7 @@ from easydev.setting import ID_EXT, SRV_JOB, DESKTOP, TOOLKIT
 
 
 class EasyDev(unohelper.Base, XServiceName, XServiceInfo,
-    Debug, Tools, Arrays, EMail, LOCalc, LOWriter, LODialog, LODataBase):
+    Debug, Tools, Arrays, EMail, LOChart, LOCalc, LOWriter, LODialog, LODataBase):
 
     service_name = implementation_name = ID_EXT
 
@@ -27,6 +28,7 @@ class EasyDev(unohelper.Base, XServiceName, XServiceInfo,
         Tools.__init__(self, self.ctx, self.sm, self.desktop, self.toolkit)
         Arrays.__init__(self)
         EMail.__init__(self)
+        LOChart.__init__(self, self.ctx, self.sm, self.desktop, self.toolkit)
         LOCalc.__init__(self, self.ctx, self.sm, self.desktop, self.toolkit)
         LODialog.__init__(self, self.ctx, self.sm, self.desktop, self.toolkit)
         LODataBase.__init__(self, self.ctx, self.sm, self.desktop, self.toolkit)